OPIK-610 removed unnecessary interface
